Move NodeResourcesFit plugin args validation to apis/config/validation
