+++ pkg/scheduler/framework/plugins/noderesources/fit.go
@@ -22,14 +22,13 @@ import (
 	"strings"
 
 	v1 "k8s.io/api/core/v1"
-	metav1validation "k8s.io/apimachinery/pkg/apis/meta/v1/validation"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/apimachinery/pkg/util/validation/field"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	v1helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
 	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/pkg/scheduler/apis/config"
+	"k8s.io/kubernetes/pkg/scheduler/apis/config/validation"
 	"k8s.io/kubernetes/pkg/scheduler/framework"
 )
 
@@ -67,56 +66,20 @@ func (f *Fit) Name() string {
 	return FitName
 }
 
-func validateFitArgs(args config.NodeResourcesFitArgs) error {
-	var allErrs field.ErrorList
-	resPath := field.NewPath("ignoredResources")
-	for i, res := range args.IgnoredResources {
-		path := resPath.Index(i)
-		if errs := metav1validation.ValidateLabelName(res, path); len(errs) != 0 {
-			allErrs = append(allErrs, errs...)
-		}
-	}
-
-	groupPath := field.NewPath("ignoredResourceGroups")
-	for i, group := range args.IgnoredResourceGroups {
-		path := groupPath.Index(i)
-		if strings.Contains(group, "/") {
-			allErrs = append(allErrs, field.Invalid(path, group, "resource group name can't contain '/'"))
-		}
-		if errs := metav1validation.ValidateLabelName(group, path); len(errs) != 0 {
-			allErrs = append(allErrs, errs...)
-		}
-	}
-
-	if len(allErrs) == 0 {
-		return nil
-	}
-	return allErrs.ToAggregate()
-}
-
 // NewFit initializes a new plugin and returns it.
 func NewFit(plArgs runtime.Object, _ framework.Handle) (framework.Plugin, error) {
-	args, err := getFitArgs(plArgs)
-	if err != nil {
-		return nil, err
+	args, ok := plArgs.(*config.NodeResourcesFitArgs)
+	if !ok {
+		return nil, fmt.Errorf("want args to be of type NodeResourcesFitArgs, got %T", plArgs)
 	}
 
-	if err := validateFitArgs(args); err != nil {
+	if err := validation.ValidateNodeResourcesFitArgs(args); err != nil {
 		return nil, err
 	}
-
 	return &Fit{
 		ignoredResources:      sets.NewString(args.IgnoredResources...),
 		ignoredResourceGroups: sets.NewString(args.IgnoredResourceGroups...),
 	}, nil
-}
-
-func getFitArgs(obj runtime.Object) (config.NodeResourcesFitArgs, error) {
-	ptr, ok := obj.(*config.NodeResourcesFitArgs)
-	if !ok {
-		return config.NodeResourcesFitArgs{}, fmt.Errorf("want args to be of type NodeResourcesFitArgs, got %T", obj)
-	}
-	return *ptr, nil
 }
 
 // computePodResourceRequest returns a framework.Resource that covers the largest
